[REFACTOR] Apply/Team Entity UserService cod move to Each Dict
