+++ src/main/java/com/PuzzleU/Server/team/dto/ApplyTeamListDto.java
@@ -11,9 +11,9 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 // 마감빠른 순 , 마감 느린 순, 인기 순, 팀빌 순
-public class TeamListDto {
+public class ApplyTeamListDto {
 
-    private List<TeamAbstractBaseDto> teamList;
+    private List<TeamAbstractDto> teamList;
     private int totalTeam;
     private int pageNo;
     private int pageSize;
