+++ src/main/java/com/PuzzleU/Server/team/dto/TeamAbstractBaseDto.java
@@ -1,0 +1,22 @@
+package com.PuzzleU.Server.team.dto;
+
+import lombok.*;
+
+import java.util.List;
+
+@Getter
+@Data
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TeamAbstractBaseDto {
+    private Long teamId;
+    private String teamTitle;
+    private String teamWriter;
+    private List<String> positionList;
+    private Integer teamNowMember;
+    private Integer teamNeed;
+    private Boolean teamUntact;
+    private List<String> teamLocations;
+    private String teamPoster;
+}
