+++ src/main/java/com/PuzzleU/Server/team/service/TeamService.java
@@ -1,5 +1,6 @@
 package com.PuzzleU.Server.team.service;
 
+import com.PuzzleU.Server.apply.repository.ApplyRepository;
 import com.PuzzleU.Server.common.api.ApiResponseDto;
 import com.PuzzleU.Server.common.api.ErrorResponse;
 import com.PuzzleU.Server.common.api.ResponseUtils;
@@ -16,7 +17,7 @@ import com.PuzzleU.Server.relations.entity.TeamLocationRelation;
 import com.PuzzleU.Server.relations.entity.TeamUserRelation;
 import com.PuzzleU.Server.relations.repository.TeamLocationRelationRepository;
 import com.PuzzleU.Server.relations.repository.TeamUserRepository;
-import com.PuzzleU.Server.team.dto.TeamCreateDto;
+import com.PuzzleU.Server.team.dto.*;
 import com.PuzzleU.Server.friendship.repository.FriendshipRepository;
 import com.PuzzleU.Server.team.entity.Team;
 import com.PuzzleU.Server.team.repository.TeamRepository;
@@ -37,6 +38,7 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -49,7 +51,7 @@ public class TeamService {
     private final TeamUserRepository teamUserRepository;
     private final TeamLocationRelationRepository teamLocationRelationRepository;
     private final PositionRepository positionRepository;
-
+    private final ApplyRepository applyRepository;
     // 팀 구인글을 등록하는 API
     @Transactional
     public ApiResponseDto<SuccessResponse> teamcreate(
@@ -313,5 +315,146 @@ public class TeamService {
         friendShipSearchResponseDto.setPageNo(pageNo);
         friendShipSearchResponseDto.setTotalElements(friendShips.getTotalElements());
      return ResponseUtils.ok(friendShipSearchResponseDto, null)  ;
+    }
+    public ApiResponseDto<TeamApplyDto> getTeamApplyTotal(UserDetails loginUser) {
+        User user = userRepository.findByUsername(loginUser.getUsername())
+                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
+        TeamApplyDto teamApplyDto = new TeamApplyDto();
+
+        for (int i = 0; i <= 1; i++) {
+            List<Team> team1;
+            Team team2;
+            if (i == 0) {
+                team1 = teamUserRepository.findByUserAndWaitingOne(user);
+                if (team1.isEmpty()) {
+                    team2 = new Team(); // 빈 팀 생성
+                } else {
+                    team2 = team1.get(0);
+                }
+            } else {
+                team1 = teamUserRepository.findByUserAndEndOne(user);
+                if (team1.isEmpty()) {
+                    team2 = new Team(); // 빈 팀 생성
+                } else {
+                    team2 = team1.get(0);
+                }
+            }
+            TeamAbstractBaseDto teamAbstractDto1 = new TeamAbstractBaseDto();
+            if (!team1.isEmpty()) {
+                List<TeamUserRelation> teamUserRelation1 = teamUserRepository.findByTeam(team2);
+                for (TeamUserRelation teamuserRelation : teamUserRelation1) {
+                    if (teamuserRelation.getIsWriter()) {
+                        teamAbstractDto1.setTeamWriter(teamuserRelation.getUser().getUserKoreaName());
+                        break;
+                    }
+                }
+                List<TeamLocationRelation> teamLocationRelation = teamLocationRelationRepository.findByTeam(team2);
+                List<String> locationList1 = new ArrayList<>();
+                for (TeamLocationRelation teamLocationRelation1 : teamLocationRelation) {
+                    String location = teamLocationRelation1.getLocation().getLocationName();
+                    locationList1.add(location);
+                }
+                teamAbstractDto1.setTeamId(team2.getTeamId());
+                teamAbstractDto1.setTeamNeed(team2.getTeamMemberNeed());
+                teamAbstractDto1.setTeamNowMember(team2.getTeamMemberNow());
+                teamAbstractDto1.setTeamTitle(team2.getTeamTitle());
+                teamAbstractDto1.setTeamPoster(team2.getCompetition().getCompetitionPoster());
+                teamAbstractDto1.setTeamLocations(locationList1);
+                List<String> PositionList1 = new ArrayList<>();
+                for (Position position : team2.getPositionList()) {
+                    PositionList1.add(position.getPositionName());
+                }
+                teamAbstractDto1.setPositionList(PositionList1);
+            }
+            if (i == 0) {
+                teamApplyDto.setTeamAbstractDtoWaiting(teamAbstractDto1);
+            } else {
+                teamApplyDto.setTeamAbstractDtoEnd(teamAbstractDto1);
+            }
+        }
+        return ResponseUtils.ok(teamApplyDto, null);
+    }
+    public ApiResponseDto<TeamListDto> getTeamApplyType(UserDetails loginUser, int pageNo, int pageSize, String sortBy, String type) {
+        User user = userRepository.findByUsername(loginUser.getUsername())
+                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
+        Pageable pageable = PageRequest.of(pageNo, pageSize, Sort.by(sortBy).descending());
+        Page<Team> teamPage;
+        if ("total".equals(type)) {
+            teamPage = new PageImpl<>(applyRepository.findByUser(user, pageable));
+        } else if ("wait".equals(type)) {
+            teamPage = new PageImpl<>(applyRepository.findByUserAndApplyStatusIsWaiting(user, pageable));
+        } else if ("end".equals(type)) {
+            teamPage = new PageImpl<>(applyRepository.findByUserAndApplyStatusIsFinished(user, pageable));
+        } else {
+            throw new RestApiException(ErrorType.NAME_NOT_PROVIDED);
+        }
+
+        TeamListDto teamListDto = new TeamListDto();
+        List<TeamAbstractBaseDto> teamAbstractDtos = teamPage.getContent().stream()
+                .map(team ->
+                {
+                    TeamAbstractBaseDto teamAbstractDto = new TeamAbstractBaseDto();
+                    List<TeamUserRelation> teamUserRelation = teamUserRepository.findByTeam(team);
+                    for (TeamUserRelation teamuserRelation : teamUserRelation) {
+                        if (teamuserRelation.getIsWriter()) {
+                            teamAbstractDto.setTeamWriter(teamuserRelation.getUser().getUserKoreaName());
+                            break;
+                        }
+                    }
+                    List<TeamLocationRelation> teamLocationRelation = teamLocationRelationRepository.findByTeam(team);
+                    List<String> locationList = new ArrayList<>();
+                    for (TeamLocationRelation teamLocationRelation1 : teamLocationRelation) {
+                        String location = teamLocationRelation1.getLocation().getLocationName();
+                        locationList.add(location);
+                    }
+                    teamAbstractDto.setTeamId(team.getTeamId());
+                    teamAbstractDto.setTeamNeed(team.getTeamMemberNeed());
+                    teamAbstractDto.setTeamNowMember(team.getTeamMemberNow());
+                    teamAbstractDto.setTeamTitle(team.getTeamTitle());
+                    teamAbstractDto.setTeamPoster(team.getCompetition().getCompetitionPoster());
+                    teamAbstractDto.setTeamLocations(locationList);
+                    List<String> PositionList = new ArrayList<>();
+                    for (Position position : team.getPositionList()) {
+                        PositionList.add(position.getPositionName());
+                    }
+                    teamAbstractDto.setPositionList(PositionList);
+                    return teamAbstractDto;
+                })
+                .collect(Collectors.toList());
+
+        teamListDto.setTotalTeam(teamAbstractDtos.size());
+        teamListDto.setTeamList(teamAbstractDtos);
+        teamListDto.setLast(teamPage.isLast());
+        teamListDto.setTotalPages(teamPage.getTotalPages());
+        teamListDto.setTotalElements(teamPage.getTotalElements());
+        teamListDto.setPageNo(pageNo);
+        teamListDto.setPageSize(pageSize);
+
+        return ResponseUtils.ok(teamListDto, null);
+    }
+    public ApiResponseDto<SuccessResponse> teamStatus(UserDetails loginUser, Long team_id, TeamStatusDto teamStatusDto)
+    {
+        Optional<Team> teamOptional  = teamRepository.findById(team_id);
+        Team team = teamOptional.orElseThrow(
+                () -> new RestApiException(ErrorType.NOT_FOUND_TEAM)
+        );
+        User user = userRepository.findByUsername(loginUser.getUsername())
+                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
+        // 유저가 권한이 없거나
+        Optional<TeamUserRelation> teamUserRelationOptional = teamUserRepository.findByUserAndTeam(user, team);
+        TeamUserRelation teamUserRelation = teamUserRelationOptional.orElseThrow(
+                () -> new RestApiException(ErrorType.NOT_FOUND_USER_TEAM)
+        );
+        if(teamUserRelation.getIsWriter()==true)
+        {
+            team.setTeamStatus(teamStatusDto.getTeamStatus());
+            teamRepository.save(team);
+            return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "팀 공고글 상태 변경완료"),null);
+        }
+        else
+        {
+            return ResponseUtils.error(ErrorResponse.of(HttpStatus.NOT_ACCEPTABLE,"권한이 없습니다"));
+        }
+
     }
 }
